rename: components -> custom-component
